app/requirements.txt
bigframes == 1.22.0
google-cloud-aiplatform == 1.70.0
google-cloud-bigquery == 3.27.0
google-cloud-pubsub == 2.26.0
google-cloud-storage == 2.18.2
gunicorn == 23.0.0
json_repair == 0.30.0
langchain == 0.2.35
langchain_core == 0.3.10
langgraph == 0.2.35
mesop == 0.12.7
protobuf == 5.28.2
vertexai == 1.70.0
json_repair == 0.30.0
langchain_google_vertexai == 2.0.8
pydantic == 2.9.1
magika == 0.5.1
